# ShopSphere-Backend/src/main/java/com/shopsphere/dto/LoginResponse.java
package com.shopsphere.dto;

import lombok.*;

@Getter
@Setter
@NoArgsConstructor
@AllArgsConstructor
@Builder
public class LoginResponse {

    private Long id;
    private String name;
    private String email;
    private String role;
    private String token;
}