# openapi.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://127.0.0.1:8080","description":"Generated server url"}],"paths":{"/api/products/{id}":{"get":{"tags":["product-controller"],"operationId":"getProductById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProductDTO"}}}}}},"put":{"tags":["product-controller"],"operationId":"updateProduct","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProductDTO"}}}}}},"delete":{"tags":["product-controller"],"operationId":"deleteProduct","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/cart/items/{cartItemId}":{"put":{"tags":["cart-controller"],"operationId":"updateCartItem","parameters":[{"name":"cartItemId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"quantity","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseCartResponse"}}}}}},"delete":{"tags":["cart-controller"],"operationId":"removeCartItem","parameters":[{"name":"cartItemId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseCartResponse"}}}}}}},"/api/products":{"get":{"tags":["product-controller"],"operationId":"getAllProducts","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}},{"name":"sortBy","in":"query","required":false,"schema":{"type":"string","default":"id"}},{"name":"direction","in":"query","required":false,"schema":{"type":"string","default":"asc"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponsePageProductDTO"}}}}}},"post":{"tags":["product-controller"],"operationId":"createProduct","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseProductDTO"}}}}}}},"/api/payments/create/{orderId}":{"post":{"tags":["payment-controller"],"operationId":"createPaymentOrder","parameters":[{"name":"orderId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponsePaymentOrderResponse"}}}}}}},"/api/orders":{"get":{"tags":["order-controller"],"operationId":"getMyOrders","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListOrderResponse"}}}}}},"post":{"tags":["order-controller"],"operationId":"createOrder","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseOrderResponse"}}}}}}},"/api/cart/items":{"post":{"tags":["cart-controller"],"operationId":"addToCart","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AddToCartRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseCartResponse"}}}}}}},"/api/auth/register":{"post":{"tags":["auth-controller"],"operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseVoid"}}}}}}},"/api/auth/login":{"post":{"tags":["auth-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseLoginResponse"}}}}}}},"/api/products/search":{"get":{"tags":["product-controller"],"operationId":"searchProducts","parameters":[{"name":"keyword","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListProductDTO"}}}}}}},"/api/products/category/{category}":{"get":{"tags":["product-controller"],"operationId":"getProductsByCategory","parameters":[{"name":"category","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListProductDTO"}}}}}}},"/api/orders/{orderId}":{"get":{"tags":["order-controller"],"operationId":"getMyOrder","parameters":[{"name":"orderId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseOrderResponse"}}}}}}},"/api/cart":{"get":{"tags":["cart-controller"],"operationId":"getCart","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseCartResponse"}}}}}}}},"components":{"schemas":{"ProductDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string","minLength":1},"description":{"type":"string","minLength":1},"price":{"type":"number","exclusiveMinimum":0},"stock":{"type":"integer","format":"int32","minimum":0},"category":{"type":"string","minLength":1},"imageUrl":{"type":"string"}},"required":["category","description","name","price","stock"]},"ApiResponseCartResponse":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/CartResponse"}}},"CartItemResponse":{"type":"object","properties":{"cartItemId":{"type":"integer","format":"int64"},"productId":{"type":"integer","format":"int64"},"productName":{"type":"string"},"price":{"type":"number"},"imageUrl":{"type":"string"},"quantity":{"type":"integer","format":"int32"},"subtotal":{"type":"number"}}},"CartResponse":{"type":"object","properties":{"cartId":{"type":"integer","format":"int64"},"items":{"type":"array","items":{"$ref":"#/components/schemas/CartItemResponse"}},"totalAmount":{"type":"number"}}},"ApiResponseProductDTO":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/ProductDTO"}}},"ApiResponsePaymentOrderResponse":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/PaymentOrderResponse"}}},"PaymentOrderResponse":{"type":"object","properties":{"orderId":{"type":"integer","format":"int64"},"razorpayOrderId":{"type":"string"},"amount":{"type":"number"},"currency":{"type":"string"},"keyId":{"type":"string"}}},"ApiResponseOrderResponse":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/OrderResponse"}}},"OrderItemResponse":{"type":"object","properties":{"productId":{"type":"integer","format":"int64"},"productName":{"type":"string"},"price":{"type":"number"},"quantity":{"type":"integer","format":"int32"},"subtotal":{"type":"number"}}},"OrderResponse":{"type":"object","properties":{"orderId":{"type":"integer","format":"int64"},"totalAmount":{"type":"number"},"status":{"type":"string","enum":["PENDING","CONFIRMED","SHIPPED","DELIVERED","CANCELLED"]},"createdAt":{"type":"string","format":"date-time"},"items":{"type":"array","items":{"$ref":"#/components/schemas/OrderItemResponse"}}}},"AddToCartRequest":{"type":"object","properties":{"productId":{"type":"integer","format":"int64"},"quantity":{"type":"integer","format":"int32","minimum":1}},"required":["productId","quantity"]},"RegisterRequest":{"type":"object","properties":{"name":{"type":"string","minLength":1},"email":{"type":"string","format":"email","minLength":1},"password":{"type":"string","maxLength":2147483647,"minLength":6}},"required":["email","name","password"]},"ApiResponseVoid":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"string"},"data":{}}},"LoginRequest":{"type":"object","properties":{"email":{"type":"string","format":"email","minLength":1},"password":{"type":"string","minLength":1}},"required":["email","password"]},"ApiResponseLoginResponse":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/LoginResponse"}}},"LoginResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"email":{"type":"string"},"role":{"type":"string"},"token":{"type":"string"}}},"ApiResponsePageProductDTO":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/PageProductDTO"}}},"PageProductDTO":{"type":"object","properties":{"totalPages":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/ProductDTO"}},"number":{"type":"integer","format":"int32"},"sort":{"$ref":"#/components/schemas/SortObject"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"numberOfElements":{"type":"integer","format":"int32"},"first":{"type":"boolean"},"last":{"type":"boolean"},"empty":{"type":"boolean"}}},"PageableObject":{"type":"object","properties":{"offset":{"type":"integer","format":"int64"},"paged":{"type":"boolean"},"pageNumber":{"type":"integer","format":"int32"},"pageSize":{"type":"integer","format":"int32"},"sort":{"$ref":"#/components/schemas/SortObject"},"unpaged":{"type":"boolean"}}},"SortObject":{"type":"object","properties":{"empty":{"type":"boolean"},"sorted":{"type":"boolean"},"unsorted":{"type":"boolean"}}},"ApiResponseListProductDTO":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"string"},"data":{"type":"array","items":{"$ref":"#/components/schemas/ProductDTO"}}}},"ApiResponseListOrderResponse":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"string"},"data":{"type":"array","items":{"$ref":"#/components/schemas/OrderResponse"}}}}}}}